Updated title screen to accept user input and pass to main scene. Renamed files for clarity
