+++ PirateGame/public/js/scenes/index.js
@@ -1,0 +1,24 @@
+import { StartScene } from "./start-scene.js";
+import { MainScene} from "./main-scene.js";
+
+const parent = document.getElementById('game-container');
+
+// Set up game 
+const config = {
+    type: Phaser.AUTO,
+    width: window.innerWidth,
+    height: window.innerHeight,
+    backgroundColor: '#2d80c9',
+    parent: parent,
+    scene: [StartScene, MainScene],     // Add all scenes in
+
+    physics: {
+        default: 'matter',  // for complex physics
+        matter: {
+            gravity: { x: 0, y: 0 },
+            debug: true
+        }
+    },
+};
+
+const game = new Phaser.Game(config);
