+++ PirateGame/public/js/scenes/start-scene.js
@@ -8,7 +8,6 @@ export class StartScene extends Phaser.Scene {
         this.background = null;
     }
 
-    // TODO accept usernames in title screen
 
     preload() {
         this.load.image('background', 'assets/water.png');
@@ -21,13 +20,12 @@ export class StartScene extends Phaser.Scene {
         const centerX = this.scale.width / 2;
         const centerY = this.scale.height / 2;
 
-        let textPressed;
 
         // Add the animated background
         this.background = this.add.tileSprite(centerX, centerY, this.scale.width, this.scale.height, 'background');
 
         const title = this.add.image(centerX, 200, 'title');
-        const ship = this.add.sprite(centerX, 400, 'ship');
+        const ship = this.add.sprite(centerX, 700, 'ship');
 
         // Animate the ship sprite
         this.anims.create({
@@ -48,20 +46,29 @@ export class StartScene extends Phaser.Scene {
             loop: -1
         });
 
+        // Get the username from the textinput
 
+        let startBtn = document.getElementById('start-btn');
+        startBtn.addEventListener('click', () => {
+            let inputText =  document.getElementById('input-text')
+            // @ts-ignore because getElementById returns HTMLElement, not HTMLInputElement- it works regardless
+            let username = inputText.value; 
 
-        // startButtonBg.on('pointerdown', () => {
-        //     // Hide the form
-        //     document.getElementById('input-form').style.display = 'none';
+            if (username) // not null or empty
+            {
+                // Hide the form before continuing
+                document.getElementById('input-form').style.display = 'none';
+                this.scene.start('MainScene', { username });      // Pass the username to the main scene
+                
+            }
+        })
 
-
-        //     this.scene.start('MainScene');      // Direct to the main game when pressed
-        // });
     }
 
 
     update() {
         // Vertically scroll the background image
+
         if (this.background) {
             this.background.tilePositionY -= 6;
         }
