+++ PirateGame/public/styles.css
@@ -25,7 +25,7 @@ body {
     align-items: center;
     position: fixed;
     gap: 12px;
-    top: 50%;
+    top: 45%;
     left: 50%;
     transform: translate(-50%, -50%);
     z-index: 9999;
@@ -63,8 +63,8 @@ body {
     font-weight: bold;
     font-size: 28px;
     text-align: center;
-    border:none;
-    background-color: rgba(13,47,79,0.9);
+    border: none;
+    background-color: rgba(13, 47, 79, 0.9);
 
     padding: 0 12px;
 
@@ -73,9 +73,7 @@ body {
 }
 
 input {
-    cursor: text;
-    user-select: all;
-    -moz-user-select: all;
+    cursor: text
 }
 
 button {
